Updated to ROS Noetic, ubuntu Focal, and Python3. Adjusted enough to start the complete system and confirm battery voltage and current is displayed by the UI.
