+++ .gitignore
@@ -1,1 +1,1 @@
-servo_settings.py
+__pycache__
