+++ nodes/roboquest_core/servo_controller.py
@@ -1,4 +1,4 @@
-#! /usr/bin/env python
+#!/usr/bin/env python3
 #
 # servo_controller.py
 # Cory Duce cory.duce@gmail.com
