+++ nodes/roboquest_core/servo_settings.py
@@ -1,0 +1,17 @@
+#
+# servo_settings.py
+# Cory Duce cory.duce@gmail.com
+# Nov 24 2020
+# 
+# v1.0 -Initial release
+#
+
+#User adjustable parameters for the 16 servo channels
+#
+#
+#                               0         1     2     3     4     5     6     7     8     9     10    11    12    13    14    15
+servo_max_angle  = [180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180,  180  ]
+servo_min_pulse  = [600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600,  600  ]
+servo_max_pulse  = [2400, 2500, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400, 2400 ]
+servo_init_angle = [90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90,   90   ]
+servo_init_delay = [0,    1,    0.3,  0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05, 0.05 ]
