+++ package.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<package format="2">
+  <name>roboquest_core</name>
+  <version>0.0.1</version>
+  <description>The roboquest_core package</description>
+
+  <maintainer email="cory.duce@gmail.com">Cory Duce</maintainer>
+  <author email="cory.duce@gmail.com">Cory Duce</author> -->
+
+  <license>GPLv2</license>
+
+  <buildtool_depend>catkin</buildtool_depend>
+  <build_depend>rospy</build_depend>
+  <build_export_depend>rospy</build_export_depend>
+  <exec_depend>rospy</exec_depend>
+
+  <export>
+  </export>
+</package>
