--- roboquest_core.launch
@@ -1,4 +1,0 @@
-<launch>
-	<node name="servo_controller" pkg="roboquest_core" type="servo_controller.py" respawn="true" />
-	<node name="motor_controller" pkg="roboquest_core" type="motor_controller.py" respawn="true" />
-</launch>
